+++ src/userAttrCache.h
@@ -176,12 +176,12 @@ public:
      * be unregistered immediately after that. If \false, the callback will be called
      * every time the attribute changes on the server and the new value is fetched.
      */
-    Handle getAttr(uint64_t user, uint8_t attrType, void* userp,
+    Handle getAttr(uint64_t user, unsigned attrType, void* userp,
                              UserAttrReqCbFunc cb, bool oneShot=false, bool fetch = true, uint64_t ph = Id::inval());
     /** @brief A promise-based version of \c getAttr. The request
      * is implicitly one-shot, as a promise can be resolved only once.
      */
-    promise::Promise<Buffer*> getAttr(uint64_t user, uint8_t attrType, uint64_t ph = Id::inval());
+    promise::Promise<Buffer*> getAttr(uint64_t user, unsigned attrType, uint64_t ph = Id::inval());
     /** @brief Unregisters an attribute request/subsequent callbacks.
      * It can be a not-yet-fetched single shot request as well. Use this method
      * to unsubscribe from further calling the corresponding callback.
